+++ elle/src/elle/Buffer.hh
@@ -64,10 +64,10 @@ namespace elle
     Buffer(int size);
     /// An uninitialized buffer of the specified size.
     explicit
-    Buffer(uint64_t size);
+    Buffer(unsigned int size);
     /// An uninitialized buffer of the specified size.
     explicit
-    Buffer(unsigned int size);
+    Buffer(unsigned long long size);
     /// A buffer containing a copy of the given data.
     Buffer(void const* data, size_t size);
     /// A buffer containing a copy of the given data.
